Rename velocity variable files
