+++ src/Makefile.am
@@ -195,8 +195,8 @@ libgrins_la_SOURCES += variables/src/variables_base.C
 libgrins_la_SOURCES += variables/src/single_variable.C
 libgrins_la_SOURCES += variables/src/fe_variables_base.C
 libgrins_la_SOURCES += variables/src/single_fe_type_variable.C
-libgrins_la_SOURCES += variables/src/primitive_flow_variables.C
-libgrins_la_SOURCES += variables/src/primitive_flow_fe_variables.C
+libgrins_la_SOURCES += variables/src/velocity_variables.C
+libgrins_la_SOURCES += variables/src/velocity_fe_variables.C
 libgrins_la_SOURCES += variables/src/displacement_variables.C
 libgrins_la_SOURCES += variables/src/displacement_fe_variables.C
 
@@ -430,8 +430,8 @@ include_HEADERS += variables/include/grins/fe_variables_base.h
 include_HEADERS += variables/include/grins/single_fe_type_variable.h
 include_HEADERS += variables/include/grins/pressure_variable.h
 include_HEADERS += variables/include/grins/pressure_fe_variable.h
-include_HEADERS += variables/include/grins/primitive_flow_variables.h
-include_HEADERS += variables/include/grins/primitive_flow_fe_variables.h
+include_HEADERS += variables/include/grins/velocity_variables.h
+include_HEADERS += variables/include/grins/velocity_fe_variables.h
 include_HEADERS += variables/include/grins/primitive_temp_variables.h
 include_HEADERS += variables/include/grins/primitive_temp_fe_variables.h
 include_HEADERS += variables/include/grins/displacement_variables.h
