+++ src/physics/include/grins/axisym_boussinesq_buoyancy.h
@@ -30,7 +30,7 @@
 #include "grins_config.h"
 #include "grins/grins_enums.h"
 #include "grins/physics.h"
-#include "grins/primitive_flow_fe_variables.h"
+#include "grins/velocity_fe_variables.h"
 #include "grins/pressure_fe_variable.h"
 #include "grins/primitive_temp_fe_variables.h"
