+++ src/physics/include/grins/boussinesq_buoyancy_base.h
@@ -28,7 +28,7 @@
 
 // GRINS
 #include "grins/physics.h"
-#include "grins/primitive_flow_fe_variables.h"
+#include "grins/velocity_fe_variables.h"
 #include "grins/pressure_fe_variable.h"
 #include "grins/primitive_temp_fe_variables.h"
