+++ src/physics/include/grins/inc_navier_stokes_stab_helper.h
@@ -28,7 +28,7 @@
 // GRINS
 #include "grins/stab_helper.h"
 #include "grins/assembly_context.h"
-#include "grins/primitive_flow_variables.h"
+#include "grins/velocity_variables.h"
 #include "grins/pressure_variable.h"
 
 // libMesh foward declarations
