+++ src/physics/include/grins/low_mach_navier_stokes_base.h
@@ -30,7 +30,7 @@
 #include "grins/physics.h"
 #include "grins/assembly_context.h"
 #include "grins/grins_enums.h"
-#include "grins/primitive_flow_fe_variables.h"
+#include "grins/velocity_fe_variables.h"
 #include "grins/pressure_fe_variable.h"
 #include "grins/primitive_temp_fe_variables.h"
 #include "grins/thermo_pressure_fe_variable.h"
